+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## Unreleased
 
+### Added
+
+- `prvhash16cw` algorithm.
+
 ## [0.44] - 2025-12-15
 
 ### Added